shared/oscar_time_jobs/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "oscar-time-jobs"
version = "0.1.0"
description = "Shared backend for the timer + alarm HERMES skills — Postgres `time_jobs` table CRUD, RRULE computation, fire dispatch."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Michael Dopp", email = "mdopp79@gmail.com" }]
dependencies = [
  "asyncpg>=0.29",
  "python-dateutil>=2.9",
  "oscar-logging",
]

[project.optional-dependencies]
test = [
  "pytest>=7",
  "pytest-asyncio>=0.23",
  "freezegun>=1.5",
]

[project.scripts]
"oscar-time-jobs" = "oscar_time_jobs.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["oscar_time_jobs*"]

[tool.setuptools.package-dir]
"" = "src"
